# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: FlowDesk API
  version: "1.0.0"
  description: |
    Multi-tenant B2B SaaS project management platform API.

    All workspace-scoped endpoints resolve `workspaceId` from the authenticated
    session — never from a client-supplied value. Access tokens expire after
    15 minutes; use `POST /auth/refresh` to obtain a new pair transparently.
servers:
  - url: http://localhost:3001/api/v1
    description: Development
  - url: https://api.flowdesk.app/v1
    description: Production
...
security:
  - BearerAuth: []
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token. Expires after 15 minutes.
  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            code:
              type: string
              description: Machine-readable error code (e.g. EMAIL_ALREADY_REGISTERED).
              example: EMAIL_ALREADY_REGISTERED
            message:
              type: string
              description: Human-readable error description.
              example: A user with that email address already exists.
            details:
              type: object
              description: Optional structured details (e.g. field-level validation errors).
              additionalProperties: true
...
    AuthTokens:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
          description: Short-lived JWT used to authenticate API requests (15 min).
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: Long-lived opaque token used to obtain a new token pair (7 days).
          example: rt_a1b2c3d4e5f6...
        expiresIn:
          type: integer
          description: Seconds until the access token expires.
          example: 900

    # -----------------------------------------------------------------------
    # Users
    # -----------------------------------------------------------------------
    UserProfile:
      type: object
      required: [id, email, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          example: 01906b5e-3a1f-7000-8c6b-d7e1f2a3b4c5
        email:
          type: string
          format: email
          example: alice@example.com
        name:
          type: string
          example: Alice Smith
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: https://cdn.flowdesk.app/avatars/alice.png
        createdAt:
          type: string
          format: date-time
          example: "2025-01-15T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-03-10T14:22:00Z"

    # -----------------------------------------------------------------------
    # Workspaces
    # -----------------------------------------------------------------------
...
  responses:
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication is required to access this resource.
...
    UnprocessableEntity:
      description: Request body failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: Request validation failed.
              details:
                email: Must be a valid email address.
                password: Must be at least 8 characters.
    InternalServerError:
      description: An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_SERVER_ERROR
              message: An unexpected error occurred. Please try again later.

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [Authentication]
      summary: Register a new user account.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:
                  type: string
                  format: email
                  example: alice@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "S3cureP@ssword!"
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: Alice Smith
      responses:
        "201":
          description: User registered successfully. Returns the new user profile and auth tokens.
          content:
            application/json:
              schema:
                type: object
                required: [user, tokens]
                properties:
                  user:
                    $ref: "#/components/schemas/UserProfile"
                  tokens:
                    $ref: "#/components/schemas/AuthTokens"
        "409":
          description: Email address is already registered to an active account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: EMAIL_ALREADY_REGISTERED
                  message: A user with that email address already exists.
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/login:
    post:
      operationId: loginUser
      tags: [Authentication]
      summary: Authenticate with email and password and receive tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: alice@example.com
                password:
                  type: string
                  format: password
                  example: "S3cureP@ssword!"
      responses:
        "200":
          description: Login successful. Returns the user profile and auth tokens.
          content:
            application/json:
              schema:
                type: object
                required: [user, tokens]
                properties:
                  user:
                    $ref: "#/components/schemas/UserProfile"
                  tokens:
                    $ref: "#/components/schemas/AuthTokens"
        "401":
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: INVALID_CREDENTIALS
                  message: The email or password you entered is incorrect.
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/refresh:
    post:
      operationId: refreshTokens
      tags: [Authentication]
      summary: Exchange a valid refresh token for a new token pair.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  description: The opaque refresh token issued at login or a previous refresh.
                  example: rt_a1b2c3d4e5f6...
      responses:
        "200":
          description: Tokens refreshed successfully.
          content:
            application/json:
              schema:
                type: object
                required: [tokens]
                properties:
                  tokens:
                    $ref: "#/components/schemas/AuthTokens"
        "401":
          description: Refresh token is invalid, expired, or has been revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: INVALID_REFRESH_TOKEN
                  message: The refresh token is invalid or has expired. Please log in again.
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/logout:
    post:
      operationId: logoutUser
      tags: [Authentication]
      summary: Revoke the current refresh token and invalidate the session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  description: The refresh token to revoke.
                  example: rt_a1b2c3d4e5f6...
      responses:
        "204":
          description: Session terminated. The refresh token has been revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...